+++ demo/cypress/example/mock-assets.js
